[APPC-1248] 1- Changed code to use kotlin functions. 2 - Moved transactions navigator and interactor to the correct package
