+++ app/src/main/java/com/asfoundation/wallet/navigator/TransactionViewNavigator.kt
@@ -1,4 +1,4 @@
-package com.asfoundation.wallet.di
+package com.asfoundation.wallet.navigator
 
 import android.content.Context
 import android.content.Intent
